pyproject.toml
[project]
name = "mcp-learn"
version = "0.1.0"
description = "MCP learn"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "autogen>=0.5.3",
    "autogen-agentchat>=0.5.3",
    "autogen-ext[azure,mcp,openai]>=0.5.3",
    "chainlit>=2.5.5",
    "colorama>=0.4.6",
    "fastmcp>=2.2.0",
    "google-adk>=0.2.0",
    "langchain>=0.3.23",
    "langchain-mcp-adapters>=0.0.9",
    "langchain-openai>=0.3.14",
    "langgraph>=0.3.31",
    "langsmith>=0.3.32",
    "litellm>=1.67.0.post1",
    "mcp>=1.6.0",
    "mlflow>=2.21.3",
    "newsapi-python>=0.2.7",
    "openai>=1.75.0",
    "openai-agents>=0.0.11",
    "pydantic-ai-slim[mcp,openai]>=0.1.3",
    "pytz>=2025.2",
    "streamlit>=1.44.1",
    "tenacity>=9.1.2",
    "tiktoken>=0.9.0",
    "yfinance>=0.2.55",
]
